Add scan scaffolding
