+++ kbfstool/md_check.go
@@ -1,0 +1,55 @@
+package main
+
+import (
+	"flag"
+	"fmt"
+
+	"github.com/keybase/kbfs/libkbfs"
+	"golang.org/x/net/context"
+)
+
+const mdCheckUsageStr = `Usage:
+  kbfstool md check input [inputs...]
+
+Each input must be in the same format as in md dump.
+
+`
+
+func mdCheckOne(ctx context.Context, config libkbfs.Config,
+	rmd libkbfs.ImmutableRootMetadata) error {
+	panic("Not implemented")
+}
+
+func mdCheck(ctx context.Context, config libkbfs.Config, args []string) (exitStatus int) {
+	flags := flag.NewFlagSet("kbfs md check", flag.ContinueOnError)
+	flags.Parse(args)
+
+	inputs := flags.Args()
+	if len(inputs) < 1 {
+		fmt.Print(mdCheckUsageStr)
+		return 1
+	}
+
+	for _, input := range inputs {
+		irmd, err := mdParseAndGet(ctx, config, input)
+		if err != nil {
+			printError("md check", err)
+			return 1
+		}
+
+		if irmd == (libkbfs.ImmutableRootMetadata{}) {
+			fmt.Printf("No result found for %q\n\n", input)
+			continue
+		}
+
+		err = mdCheckOne(ctx, config, irmd)
+		if err != nil {
+			printError("md check", err)
+			return 1
+		}
+
+		fmt.Print("\n")
+	}
+
+	return 0
+}
